// frontend/src/sections/customer-returns/preferences.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type CustomerReturnPreferencesVariables = {
  storeId: string;
};

export type CustomerReturnPreferencesResult = {
  preferences: {
  disableManualReturns: boolean;
  invoiceStatusOptions: Array<"NEW" | "ALLOCATED" | "PICKED" | "SHIPPED" | "DELIVERED" | "RECEIVED" | "VERIFIED" | "CANCELLED">;
};
};

export const CustomerReturnPreferences = {
  query: "query customerReturnPreferences($storeId: String!) {\n  preferences(storeId: $storeId) {\n    disableManualReturns\n    invoiceStatusOptions\n  }\n}",
} as TypedDocument<CustomerReturnPreferencesResult, CustomerReturnPreferencesVariables>;